fixed issue with using arbitrary cartesian communicator
